# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ilyas Store eCommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get user cart items
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItemResponse"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemRequest"
      responses:
        "201":
          description: Cart item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItemResponse"
  /cart/{id}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
              required: [quantity]
      responses:
        "200":
          description: Cart item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItemResponse"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove cart item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ORDERS
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Banner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "200":
          description: Banner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # COUPONS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # USER
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, admin]
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, isBlocked, createdAt]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, productCount, createdAt]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        image:
          type: string
      required: [name]

    # PRODUCT
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, userName, rating, createdAt]
...
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        type:
          type: string
        value:
          type: string
        price:
          type: number
        stock:
          type: integer
        sku:
          type: string
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, productId, type, value, price, stock]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        brand:
          type: string
        basePrice:
          type: number
          nullable: true
        costPrice:
          type: number
          nullable: true
        compareAtPrice:
          type: number
          nullable: true
        description:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        isFeatured:
          type: boolean
        categoryId:
          type: integer
        categoryName:
          type: string
        averageRating:
          type: number
        reviewCount:
          type: integer
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        createdAt:
          type: string
          format: date-time
      required: [id, name, stock, isFeatured, categoryId, createdAt]
...
    CartItemRequest:
      type: object
      properties:
        productId:
          type: integer
        variantId:
          type: integer
          nullable: true
        quantity:
          type: integer
      required: [productId, quantity]
    CartItemResponse:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        variantId:
          type: integer
          nullable: true
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
      required: [id, productId, quantity, product]

    # BANNER
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
        image:
          type: string
        link:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, image, isActive, sortOrder, createdAt]
    CreateBannerBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        image:
          type: string
        link:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required: [title, image]

    # COUPON
...